--- a/benchmark/wer_benchmark.py
+++ b/benchmark/wer_benchmark.py
@@ -1,6 +1,7 @@
 import argparse
 import json
 import os
+
 from datasets import load_dataset
 from evaluate import load
 from tqdm import tqdm