--- a/faster_whisper/audio.py
+++ b/faster_whisper/audio.py
@@ -1,7 +1,7 @@
 from typing import BinaryIO, Union
 
-import torchaudio
 import torch
+import torchaudio
 
 
 def decode_audio(
@@ -45,7 +45,13 @@ def pad_or_trim(array, length: int, *, axis: int = -1):
         return array[idx]
 
     if array.shape[axis] < length:
-        pad_widths = [0,] * array.ndim * 2
+        pad_widths = (
+            [
+                0,
+            ]
+            * array.ndim
+            * 2
+        )
         pad_widths[2 * axis] = length - array.shape[axis]
         array = torch.nn.functional.pad(array, tuple(pad_widths[::-1]))