--- a/requirements.txt
+++ b/requirements.txt
@@ -7,4 +7,3 @@ transformers
 pyannote-audio>=3.1.1
 torch>=2.1.1 
 torchaudio>=2.1.2
-jsons>=1.6.3