… faster_whisper/vad.py
… tests/test_transcribe.py