* removing the need for `jsons` dependency
* remove tokenizer reinitialization when language or task is changed since it can be changed directly without creating a new tokenizer
* remove the need for a separate `encode_batched` function
* enable word timestamps using original functions
* remove `PyAV` and use `torchaudio` instead, this fixes the memory leak in the resampler
* use gpu in feature extraction for 35x speedup, I also removed the old feature extraction since it has no use case, so there's no need to keep it while we have a faster one that has the same accuracy

the diff seems to be messed up for a reason, but the changes are minimal