@trungkienbkhn 2024-06-25
When testing, I encountered problem with word timestamps, they are a bit weird:
[68.42s -> 68.92s]  => [41.48s -> 42.58s] => [68.92s -> 71.76s]
```
[0.00s -> 1.88s]  Would you see what you can get for this, please?
[5.36s -> 7.44s]  Not the royal ring, your highness.
[10.50s -> 10.94s]  Shh!
[12.12s -> 12.60s]  Do you want to help out?
[27.38s -> 28.44s]  Excuse me.
[45.36s -> 49.24s]  Is that man actually royalty?
[52.74s -> 54.02s]  No, madame.
[55.54s -> 57.40s]  But you called him your highness.
[59.42s -> 60.74s]  It was a faux pas.
[62.42s -> 63.34s]  Please forget about it.
[64.20s -> 67.12s]  You can trust me.
[68.42s -> 68.92s]  I won't tell.
[41.48s -> 42.58s]  Madame, I am...
[68.92s -> 71.76s]  Extraordinary man of destiny.
[92.46s -> 96.00s]  Your highness.
[100.20s -> 101.82s]  Your highness, don't be alarmed.
[105.18s -> 106.26s]  I can be trusted.
[108.82s -> 110.56s]  Are you one of my subjects?
[112.96s -> 113.60s]  No.
[114.38s -> 118.64s]  I'm an American, Fanny Eubanks of Omaha.
[121.44s -> 122.66s]  I couldn't help overhearing.
[126.06s -> 126.38s]  If you're in trouble and there's some way I can help,
[98.92s -> 101.24s]  thank you, but I cannot accept.
[105.72s -> 108.84s]  You've already risked too much just in speaking to me.
[112.22s -> 118.18s]  I still want to help.
[123.90s -> 127.10s]  You must understand, I have powerful enemies.
[131.92s -> 135.42s]  They may be watching even as we...
[140.60s -> 146.26s]  My God, you're attractive.
[121.52s -> 122.52s]  It's late.
[126.24s -> 127.54s]  I must go.
[128.92s -> 134.88s]  As he left?
[141.82s -> 143.78s]  Yes, just a moment to calm.
[148.94s -> 149.74s]  Good.
[151.90s -> 152.28s]  Please.
[153.72s -> 158.90s]  You must tell me where he lives.
[164.14s -> 166.36s]  I feel it only fair to wonder.
[168.64s -> 169.78s]  I know.
[170.68s -> 174.12s]  He told me he has powerful enemies.
[176.62s -> 178.06s]  There may also be an emotional risk.
[182.24s -> 182.56s]  You see, his highness has been a widower for five years.
[158.92s -> 160.20s]  For five years?
[174.20s -> 177.00s]  Please, your highness.
[181.04s -> 187.32s]  Fanny, the Freedom Fighters thank you.
[195.52s -> 197.20s]  This is for the overhead.
[200.52s -> 202.24s]  This goes to you, Arthur.
[206.08s -> 208.70s]  This goes to you, Andre.
[211.58s -> 212.50s]  This goes to me, which means it's time to go to Zurich.
[188.92s -> 190.18s]  Thank you.
```

I didn't use `word_timestamps=True`:
```python3
model = WhisperModel("large-v3", device="cuda")
segments, info = model.transcribe(audio_path, word_timestamps=True)
for segment in segments:
    print("[%.2fs -> %.2fs] %s" % (segment.start, segment.end, segment.text))
```
Could you take a look ?