@Jiltseb 2024-06-25
Yes, The error in timestamps looks to repeat after every batch, it starts a lot earlier than the final timestamp of the previous batch.