@MahmoudAshraf97 2024-06-25
I managed to reproduce this on several audio files but after pulling a new copy of the repo I couldn't anymore, can someone test again and provide a minimum example?