@Jiltseb 2024-06-25
Used the example `faster-whisper/tests/data/physicsworks.wav` to test both batched and sequential versions.

Batched version:

```
[0.00s -> 27.44s] Now I want to return to the conservation of mechanical energy. I have here a pendulum. I have an object that weighs 15 kilograms, and I can lift it up one meter, which I have done now. That means I've done work. MGH is the work I have done, believe me. I've increased the potential energy of this object. 15 times 10 is about 150 joules. If I let it fall,
[28.11s -> 53.91s] then that will be converted to kinetic energy. If I would let it swing from one meter height, and you would be there, and it would hit you, you'd be dead. 150 joules is enough to kill you. They use these devices. It's called a wrecker ball. They use them to demolish buildings. You lift up a very heavy object, even heavier than this,
[54.35s -> 82.93s] and then you let it go, you swing it, thereby converting gravitational potential energy into kinetic energy, and that way you can demolish a building. You just let it hit... and it breaks a building. And that's the whole idea of wrecking. So you're using, then, the conversion of gravitational potential energy to kinetic energy.
[84.21s -> 113.45s] Now, I am such a strong believer of the conservation of mechanical energy that I am willing to put my life on the line. If I release that bulb from a certain height, then that bulb can never come back to a point where the height is any larger. If I release it from this height,
[114.17s -> 137.71s] and it swings, then when it reaches here, it could not be higher. There is a conversion from gravitational potential energy to kinetic energy back to gravitational potential energy, and it will come to a stop here. And when it swings back, it should not be able to reach any higher, provided that I do not give this object an initial speed when I stand here.
[141.28s -> 169.74s] the conservation of mechanical energy for 100%. I may not trust myself. I'm going to release this object, and I hope I will be able to do it at zero speed, so that when it comes back, it may touch my chin, but it may not crush my chin. I want you to be extremely quiet, because this is no joke. If I don't succeed in giving it zero speed,
[170.18s -> 187.62s] Then, this will be my last lecture. I will close my eyes. I don't want to see this. So please be very quiet. I almost didn't sleep all night. 3, 2, 1, 0.
[200.73s -> 202.73s] Physics works, and I'm still alive.
```
Did not see timestamps issue for this test. The timestamps are still **not** sparse, are you yet to commit those changes?

Sequential version:
```
[0.00s -> 8.98s]  Now I want to return to the conservation of mechanical energy. I have here a pendulum.
[20.48s -> 25.00s]  I have an object that weighs 15 kilograms, and I can lift it up one meter, which I have
[30.22s -> 36.12s]  done now. That means I've done work. MGH is the work I have done, believe me. I've increased
[42.38s -> 48.14s]  the potential energy of this object. 15 times 10 is about 150 joules. If I let it fall,
[48.14s -> 53.56s]  I use them to demolish buildings. You lift up a very heavy object, even heavier than
[59.98s -> 66.12s]  this, and then you let it go. You swing it, thereby converting gravitational potential
[73.08s -> 82.42s]  energy into kinetic energy, and that way you can demolish a building. You just let it hit,
[93.42s -> 99.78s]  and it breaks a building. And that's the whole idea of wrecking. So you are using then the
[100.10s -> 110.40s]  that bob from a certain height, then that bob can never come back to a point where the height
[120.98s -> 128.02s]  is any larger. If I release it from this height, and it swings, then when it reaches here,
[135.30s -> 140.78s]  it could not be higher. There is a conversion from gravitational potential energy to kinetic
[146.44s -> 149.92s]  energy back to gravitational potential energy, and it will come to a stop here. And when it
[156.32s -> 163.90s]  comes back, it may touch my chin, but it may not crush my chin. I want you to be extremely
[178.48s -> 185.58s]  quiet, because this is no joke. If I don't succeed in giving it zero speed, then this
[193.10s -> 200.52s]  will be my last lecture. I will close my eyes. I don't want to see this. So please be very
[201.40s -> 201.58s]  quiet.
[230.28s -> 230.76s]  Physics works, and I'm still alive.
```
Surprisingly, the timestamps here are wrong, as it ends at 230 sec, longer than the total length of the audio.