@Jiltseb 2024-06-25
Ah I had to remove and install latest one again, it shows similar results now. I will compare with previous versions and get back.